+++ server/resource/template/fe/table.vue.tpl
@@ -110,7 +110,7 @@ import {
     find{{.StructName}},
     get{{.StructName}}List
 } from "@/api/{{.PackageName}}";  //  此处请自行替换地址
-import { formatTimeToStr } from "@/utils/data";
+import { formatTimeToStr } from "@/utils/date";
 import infoList from "@/components/mixins/infoList";
 
 export default {
