+++ web/src/view/superAdmin/dictionary/sysDictionary.vue
@@ -117,7 +117,7 @@ import {
     findSysDictionary,
     getSysDictionaryList
 } from "@/api/sysDictionary";  //  此处请自行替换地址
-import { formatTimeToStr } from "@/utils/data";
+import { formatTimeToStr } from "@/utils/date";
 import infoList from "@/components/mixins/infoList";
 export default {
   name: "SysDictionary",
