+++ web/src/view/superAdmin/dictionary/sysDictionaryDetail.vue
@@ -114,7 +114,7 @@ import {
   findSysDictionaryDetail,
   getSysDictionaryDetailList
 } from "@/api/sysDictionaryDetail"; //  此处请自行替换地址
-import { formatTimeToStr } from "@/utils/data";
+import { formatTimeToStr } from "@/utils/date";
 import infoList from "@/components/mixins/infoList";
 
 export default {
