+++ web/src/view/superAdmin/operation/sysOperationRecord.vue
@@ -113,7 +113,7 @@ import {
   getSysOperationRecordList,
   deleteSysOperationRecordByIds
 } from '@/api/sysOperationRecord' //  此处请自行替换地址
-import { formatTimeToStr } from '@/utils/data'
+import { formatTimeToStr } from '@/utils/date'
 import infoList from '@/components/mixins/infoList'
 
 export default {
